pyproject.toml
[project]
name = "edenplace-broadsheet-app"
version = "0.1.0"
description = "A Streamlit app for generating broadsheet reports from EdenPlace data"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "jinja2>=3.1.6",
    "more-itertools>=10.7.0",
    "openpyxl==3.1.5",
    "pandas==2.2.3",
    "streamlit-nightly==1.41.1.dev20241210",
]
